% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Uscore.R
\name{calculate_Uscore}
\alias{calculate_Uscore}
\title{Calculate rankings and scores for query data and given signature set}
\usage{
calculate_Uscore(
  matrix,
  features,
  maxRank = 1500,
  chunk.size = 1000,
  ncores = 1,
  w_neg = 1,
  ties.method = "average",
  storeRanks = FALSE,
  force.gc = FALSE,
  name = "_UCell"
)
}
\arguments{
\item{matrix}{the data matrix}

\item{features}{signature features}

\item{maxRank}{rank cutoff (1500)}

\item{chunk.size}{cells per matrix chunk (1000)}

\item{ncores}{number of cores to use for parallelization (1)}

\item{w_neg}{can W be negative? (1)}

\item{ties.method}{how to break ties? ("average")}

\item{storeRanks}{store ranks? (FALSE)}

\item{force.gc}{force garbage collection? (FALSE)}

\item{name}{suffix for temp objects ("_UCell")}
}
\value{
a list of results
}
\description{
Calculate rankings and scores for query data and given signature set
}
