% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genes.R
\name{check_genes}
\alias{check_genes}
\title{Check if all genes in signatures are found in data matrix
otherwise add zero counts in data-matrix to complete it}
\usage{
check_genes(matrix, features)
}
\arguments{
\item{matrix}{the data matrix}

\item{features}{features that must be present (or else completed)}
}
\value{
\preformatted{     the signature-complete data matrix
}
}
\description{
Check if all genes in signatures are found in data matrix
otherwise add zero counts in data-matrix to complete it
}
