% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_signature_names.R
\name{check_signature_names}
\alias{check_signature_names}
\title{Check signature names and add standard names if missing}
\usage{
check_signature_names(features)
}
\arguments{
\item{features}{the signature feature names}
}
\value{
\preformatted{         cleaned names
}
}
\description{
Check signature names and add standard names if missing
}
