% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_ranks_data_table.R
\name{data_to_ranks_data_table}
\alias{data_to_ranks_data_table}
\title{Calculate features' ranks from expression data matrices}
\usage{
data_to_ranks_data_table(data, ties.method = "average")
}
\arguments{
\item{data}{the expression data matrix}

\item{ties.method}{how to break ties}
}
\value{
\preformatted{             a data.table of ranks 
}
}
\description{
Calculate features' ranks from expression data matrices
}
