% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{get_gene_idx}
\alias{get_gene_idx}
\title{Get indices for genes in data matrix}
\usage{
get_gene_idx(all_genes, signature, missing_genes = "impute")
}
\arguments{
\item{all_genes}{Dictionary of all accepted gene names}

\item{signature}{A gene signature}

\item{missing_genes}{How to handle missing genes in dictionary:
"impute": impute expression to zero; "skip": remove missing
genes from signature}
}
\value{
List of indices for genes in signatures
}
\description{
Transform gene names into indices. Handle missing genes either by
imputing or by skipping them.
}
