% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{knn_smooth_scores}
\alias{knn_smooth_scores}
\title{Smoothing scores by KNN}
\usage{
knn_smooth_scores(matrix = NULL, nn = NULL)
}
\arguments{
\item{matrix}{Input data matrix}

\item{nn}{A nearest neighbor object returned by
\link[BiocNeighbors:findKNN-methods]{BiocNeighbors::findKNN}}
}
\value{
A dataframe of knn-smoothed scores
}
\description{
Smoothing scores by KNN
}
