% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings2Uscore.R
\name{rankings2Uscore}
\alias{rankings2Uscore}
\title{Get signature scores from precomputed rank matrix}
\usage{
rankings2Uscore(
  ranks_matrix,
  features,
  chunk.size = 1000,
  w_neg = 1,
  ncores = 1,
  force.gc = FALSE,
  name = "_UCell"
)
}
\arguments{
\item{ranks_matrix}{a rank matrix}

\item{features}{signature features}

\item{chunk.size}{how many cells per matrix chunk}

\item{w_neg}{can W be negative?}

\item{ncores}{how many cores to use for parallelization}

\item{force.gc}{force garbage collection to recover RAM? (FALSE)}

\item{name}{name suffix of temporary objects ("_UCell")}
}
\value{
\preformatted{               a list of results 
}
}
\description{
Get signature scores from precomputed rank matrix
}
