% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{SmoothKNN_sce}
\alias{SmoothKNN_sce}
\title{Smooth signature scores by kNN (SingleCellExperiment)}
\usage{
SmoothKNN_sce(
  obj = NULL,
  signature.names = NULL,
  reduction = "PCA",
  k = 10,
  BNPARAM = AnnoyParam(),
  suffix = "_kNN",
  sce.expname = "UCell"
)
}
\arguments{
\item{obj}{Input object stored in \link{SingleCellExperiment}}

\item{signature.names}{The names of the signatures (or any numeric metadata
column) for which to calculate kNN-smoothed scores}

\item{reduction}{Which dimensionality reduction to use for kNN smoothing.
It must be already present in the input object.}

\item{k}{Number of neighbors for kNN smoothing}

\item{BNPARAM}{A \link{BiocNeighborParam} object specifying the algorithm to use
for kNN calculation.}

\item{suffix}{The suffix to append to metadata columns for the new
knn-smoothed scores}

\item{sce.expname}{For sce objects only - which experiment stores the
signature}
}
\description{
This function performs smoothing of single-cell scores by weighted
average of the k-nearest neighbors. It can be useful to 'impute' scores by
neighboring cells and partially correct data sparsity. While this function
has been designed to smooth UCell scores, it can be applied to any numerical
metadata contained in a SingleCellExperiment object
}
\examples{
# Run UCell
library(SingleCellExperiment)
library(scater)
data(sample.matrix)
sce <- SingleCellExperiment(list(counts=sample.matrix))
gene.sets <- list( Tcell_signature = c("CD2","CD3E","CD3D"),
                  Myeloid_signature = c("SPI1","FCER1G","CSF1R"))
sce <- ScoreSignatures_UCell(sce, features=gene.sets, name=NULL)
altExp(sce, 'UCell')
# Run PCA
sce <- logNormCounts(sce)
sce <- runPCA(sce, scale=TRUE, ncomponents=20)
# Smooth signatures
sce <- SmoothKNN(sce, reduction="PCA", signature.names=names(gene.sets))
altExp(sce, 'UCell_kNN')

}
