% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.matrix.R
\name{split_data.matrix}
\alias{split_data.matrix}
\title{split data matrix into cell chunks}
\usage{
split_data.matrix(matrix, chunk.size = 1000)
}
\arguments{
\item{matrix}{the data matrix}

\item{chunk.size}{how many cells per submatrix}
}
\value{
a list of chunked matrices
}
\description{
split data matrix into cell chunks
}
