% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_stat.R
\name{u_stat}
\alias{u_stat}
\title{Calculate AUC as Mann–Whitney U statistic from a vector of ranks}
\usage{
u_stat(rank_value, maxRank = 1000, sparse = F)
}
\arguments{
\item{rank_value}{the ranks}

\item{maxRank}{limit of ranks to include}

\item{sparse}{is the vector sparse}
}
\value{
\preformatted{         AUC value for the vector
}
}
\description{
Calculate AUC as Mann–Whitney U statistic from a vector of ranks
}
