% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_stat_signature_list.R
\name{u_stat_signature_list}
\alias{u_stat_signature_list}
\title{Calculate AUC for a list of signatures, from a ranks matrix}
\usage{
u_stat_signature_list(
  sig_list,
  ranks_matrix,
  maxRank = 1000,
  sparse = F,
  w_neg = 1
)
}
\arguments{
\item{sig_list}{a list of signatures}

\item{ranks_matrix}{a matrix of ranks}

\item{maxRank}{rank cutoff, for u_stat}

\item{sparse}{is the rank matrix sparse? (FALSE)}

\item{w_neg}{can W be na}
}
\description{
Calculate AUC for a list of signatures, from a ranks matrix
}
