% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AddModuleScore_UCell}
\alias{AddModuleScore_UCell}
\title{Calculate module enrichment scores using Mann-Whitney U statistic from single-cell data (Seurat interface)}
\usage{
AddModuleScore_UCell(
  obj,
  features,
  maxRank = 1500,
  chunk.size = 1000,
  ncores = 1,
  storeRanks = F,
  assay = NULL,
  slot = "data",
  force.gc = FALSE,
  seed = 123,
  name = "_UCell"
)
}
\arguments{
\item{features}{A list of signatures, for example: \code{list( Tcell_signature = c("CD2","CD3E","CD3D"), Myeloid_signature = c("SPI1","FCER1G","CSF1R"))}}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed.
Note: this parameter is ignored if precalc.ranks are specified}

\item{chunk.size}{Number of cells to be processed simultaneously (lower size requires slightly more computation but reduces memory demands)}

\item{ncores}{Number of processors to parallelize computation. Requires package \code{future}}

\item{storeRanks}{Store ranks matrix in Seurat object (@misc slot) for fast subsequent computations. This option will demand large amounts of RAM.}

\item{assay}{Pull out data from this assay of the Seurat object (if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object}

\item{force.gc}{Explicitly call garbage collector to reduce memory footprint}

\item{seed}{Integer seed for 'future' parallel execution}

\item{name}{Name tag that will be appended at the end of each signature name, "_UCell" by default (e.g. signature score in meta data will be named: Tcell_signature_UCell)}

\item{object}{Seurat object}
}
\value{
Returns a Seurat object with module/signature enrichment scores added to object meta data; each score is stored as the corresponding signature name provided in \code{features} followed by the tag given in \code{name} (or "_UCell" by default )
}
\description{
Given a Seurat object, calculates module/signature enrichment scores on single-cell level using Mann-Whitney U statistic.
Returned scores are normalized U statistic (equivalent to AUC - Area Under the Curve).
In contrast to Seurat's AddModuleScore (based on population average gene expression binning) this score depend only on the gene expression ranks of individual cell, and therefore is robust across datasets.
}
\examples{
## Not run:
library(UCell)
SeuratObject <- AddModuleScore_UCell(SeuratObject,features=list( Tcell_signature = c("CD2","CD3E","CD3D"), Myeloid_signature = c("SPI1","FCER1G","CSF1R")))
SeuratObject$Tcell_signature_UCell
head(SeuratObject@meta.data)
## End (Not run)
}
