% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{AddModuleScore_UCell}
\alias{AddModuleScore_UCell}
\title{Calculate module enrichment scores from single-cell data (Seurat interface)}
\usage{
AddModuleScore_UCell(
  obj,
  features,
  maxRank = 1500,
  chunk.size = 1000,
  ncores = 1,
  storeRanks = F,
  assay = NULL,
  slot = "data",
  ties.method = "average",
  force.gc = FALSE,
  seed = 123,
  name = "_UCell"
)
}
\arguments{
\item{features}{A list of signatures, for example: \code{list( Tcell_signature = c("CD2","CD3E","CD3D"), Myeloid_signature = c("SPI1","FCER1G","CSF1R"))}}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed.}

\item{chunk.size}{Number of cells to be processed simultaneously (lower size requires slightly more computation but reduces memory demands)}

\item{ncores}{Number of processors to parallelize computation. Requires package \code{future}}

\item{storeRanks}{Store ranks matrix in Seurat object (@misc slot) for fast subsequent computations. This option will demand large amounts of RAM.}

\item{assay}{Pull out data from this assay of the Seurat object (if NULL, use \code{DefaultAssay(obj)})}

\item{slot}{Pull out data from this slot of the Seurat object}

\item{ties.method}{How ranking ties should be resolved (passed on to \link[data.table:frank]{data.table::frank})}

\item{force.gc}{Explicitly call garbage collector to reduce memory footprint}

\item{seed}{Integer seed for \link[future.apply:future_lapply]{future.apply::future_lapply} parallel execution}

\item{name}{Name tag that will be appended at the end of each signature name, "_UCell" by default (e.g. signature score in meta data will be named: Myeloid_signature_UCell)}

\item{object}{Seurat object}
}
\value{
Returns a Seurat object with module/signature enrichment scores added to object meta data; each score is stored as the corresponding signature name provided in \code{features} followed by the tag given in \code{name} (or "_UCell" by default )
}
\description{
Given a Seurat object, calculates module/signature enrichment scores at single-cell level using the Mann-Whitney U statistic.
UCell scores are normalized U statistics (between 0 and 1), and they are mathematically related to the Area under the ROC curve (see \href{https://doi.org/10.1256/003590002320603584}{Mason and Graham})
}
\details{
In contrast to Seurat's AddModuleScore, which is normalized by binning genes of similar expression at the population level, UCell scores depend
only on the gene expression ranks of individual cell, and therefore they are robust across datasets regardless of dataset composition.
}
\examples{
## Not run:
library(UCell)
gene.sets <- list(Tcell_signature = c("CD2","CD3E","CD3D"),
                Myeloid_signature = c("SPI1","FCER1G","CSF1R"))
SeuratObject <- AddModuleScore_UCell(SeuratObject,features = gene.sets)
SeuratObject$Tcell_signature_UCell
head(SeuratObject@meta.data)
## End (Not run)
}
