% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\docType{data}
\name{sample.matrix}
\alias{sample.matrix}
\title{Sample dataset to test UCell installation}
\format{
A sparse matrix of 600 cells and 20729 genes.
}
\usage{
sample.matrix
}
\description{
A sparse matrix (class "dgCMatrix") of single-cell transcriptomes for 600 cells and 20729 genes.
It can be used for quick tests of UCell functionalities
}
\keyword{datasets}
