% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ScoreSignatures_UCell}
\alias{ScoreSignatures_UCell}
\title{Calculate module enrichment scores from single-cell data}
\usage{
ScoreSignatures_UCell(
  matrix = NULL,
  features,
  precalc.ranks = NULL,
  maxRank = 1500,
  w_neg = 1,
  chunk.size = 1000,
  ncores = 1,
  ties.method = "average",
  force.gc = FALSE,
  seed = 123
)
}
\arguments{
\item{matrix}{A gene vs. cell data matrix, either in sparse or dense format. Leave empty if providing a rank matrix with \code{precalc.ranks}}

\item{features}{A list of signatures, for example: \code{list( Tcell_signature = c("CD2","CD3E","CD3D"), Myeloid_signature = c("SPI1","FCER1G","CSF1R"))}}

\item{precalc.ranks}{A sparse matrix of pre-calculated ranks, obtained with \code{\link{StoreRankings_UCell}}}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed.
Note: this parameter is ignored if \code{precalc.ranks} are specified}

\item{w_neg}{Weight on negative genes in signature. e.g. \code{w_neg=1} weighs equally up- and down-regulated genes, \code{w_neg=0.5} gives 50\% less importance to negative genes}

\item{chunk.size}{Number of cells to be processed simultaneously (lower size requires slightly more computation but reduces memory demands)}

\item{ncores}{Number of processors to parallelize computation. Requires package \code{future}}

\item{ties.method}{How ranking ties should be resolved (passed on to \link[data.table:frank]{data.table::frank})}

\item{force.gc}{Explicitly call garbage collector to reduce memory footprint}

\item{seed}{Integer seed for \link[future.apply:future_lapply]{future.apply::future_lapply} parallel execution}
}
\value{
Returns a dataframe of signature scores for each cell
}
\description{
Given a gene vs. cell matrix, calculates module/signature enrichment scores on single-cell level using Mann-Whitney U statistic.
UCell scores are normalized U statistics (between 0 and 1), and they are mathematically related to the Area under the ROC curve (see \href{https://doi.org/10.1256/003590002320603584}{Mason and Graham})
}
\details{
These scores only depend on the gene expression ranks of individual cell, and therefore they are robust across datasets regardless of dataset composition.
}
\examples{
## Not run:
library(UCell)
my.matrix <- UCell::sample.matrix
gene.sets <- list( Tcell_signature = c("CD2","CD3E","CD3D"),
                 Myeloid_signature = c("SPI1","FCER1G","CSF1R"))
scores <- ScoreSignatures_UCell(my.matrix, features=gene.sets)
scores[1:5,]
## End (Not run)
}
