% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothKNN.R
\name{SmoothKNN}
\alias{SmoothKNN}
\title{Smooth signature scores by KNN}
\usage{
SmoothKNN(
  obj = NULL,
  reduction = "pca",
  k = 10,
  BNPARAM = AnnoyParam(),
  signature.names = NULL,
  suffix = "_kNN"
)
}
\arguments{
\item{obj}{Input object - either a \link{SingleCellExperiment} object
or a \link{Seurat} object.}

\item{reduction}{Which dimensionality reduction to use for kNN smoothing.
It must be already present in the input object.}

\item{k}{Number of neighbors for kNN smoothing}

\item{BNPARAM}{A \link{BiocNeighborParam} object specifying the algorithm to use
for kNN calculation.}

\item{signature.names}{The names of the signatures (or any numeric metadata
column) for which to calculate kNN-smoothed scores}

\item{suffix}{The suffix to append to metadata columns for the new
knn-smoothed scores}
}
\description{
This function performs smoothing of single-cell scores by weighted
average of the k-nearest neighbors. It can be useful to 'impute' scores by
neighboring cells and partially correct data sparsity. While this function
has been designed to smooth UCell scores, it can be applied to any numerical
metadata contained in SingleCellExperiment or Seurat objects
}
\examples{
# Run UCell
library(Seurat)
gene.sets <- list(Tcell = c("CD2","CD3E","CD3D"),
                Myeloid = c("SPI1","FCER1G","CSF1R"))
data(sample.matrix)
obj <- Seurat::CreateSeuratObject(sample.matrix)                

obj <- AddModuleScore_UCell(obj,features = gene.sets, name=NULL)
# Run PCA
obj <- FindVariableFeatures(obj) |> ScaleData() |> RunPCA()
# Smooth signatures
obj <- SmoothKNN(obj, reduction="pca", signature.names=names(gene.sets))
head(obj[[]])

}
