% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{StoreRankings_UCell}
\alias{StoreRankings_UCell}
\title{Calculate and store gene rankings for a single-cell dataset}
\usage{
StoreRankings_UCell(
  matrix,
  maxRank = 1500,
  chunk.size = 1000,
  ncores = 1,
  force.gc = FALSE,
  seed = 123
)
}
\arguments{
\item{matrix}{A gene vs. cell data matrix, either in sparse or dense format}

\item{maxRank}{Maximum number of genes to rank per cell; above this rank, a given gene is considered as not expressed}

\item{chunk.size}{Number of cells to be processed simultaneously (lower size requires slightly more computation but reduces memory demands)}

\item{ncores}{Number of processors to parallelize computation. Requires package \code{future}}

\item{force.gc}{Explicitly call garbage collector to reduce memory footprint}

\item{seed}{Integer seed for 'future' parallel execution}
}
\value{
Returns a sparse matrix of pre-calculated ranks that can be used multiple times to evaluate different signatures
}
\description{
Given a gene vs. cell matrix, calculates the rankings of expression for all genes in each cell. It can then be applied to the
function \code{\link{ScoreSignatures_UCell}} to evaluate gene signatures on the gene expression ranks of individual cells.
}
\examples{
## Not run:
library(UCell)
ranks <- StoreRankings_UCell(expression_matrix)
ranks[1:5,1:5]
scores <- ScoreSignatures_UCell(expression_matrix,features=feature.list, precalc.ranks=ranks)
## End (Not run)
}
